+++ Assets/Concept 2.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dae850f0bf66917488b348eeb6d3300d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
