+++ Assets/Scripts/Player2D.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a9223adb1e3079449b9f099263d35c15
+guid: d28a231737df2ca4093312a117c714e5
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
